requirements.txt
absl-py==0.13.0
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
argon2-cffi==20.1.0
astroid==2.5.6
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
btrfsutil==5.12.1
CacheControl==0.12.6
cachetools==4.2.2
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.7
cupshelpers==1.0
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
distro==1.5.0
dnspython==1.16.0
docopt==0.6.2
entrypoints==0.3
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
google-auth==1.31.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
idna==3.2
importlib-metadata==4.0.1
ipykernel==5.5.3
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyutils==0.6
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
LibAppArmor==3.0.1
louis==3.18.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mediapipe==0.8.5
mistune==0.8.4
more-itertools==8.7.0
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netsnmp-python==1.0a1
nltk==3.6.2
notebook==6.3.0
npyscreen==4.10.5
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.2.54
opencv-python==4.5.2.54
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.21.2
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
ply==3.11
progress==1.5
prometheus-client==0.10.1
prompt-toolkit==3.0.18
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
Pygments==2.9.0
PyGObject==3.40.1
pylint==2.8.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyrsistent==0.17.3
pysmbc==1.0.23
python-dateutil==2.8.1
python-libtorrent==1.2.11
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
rados==2.0.0
rbd==2.0.0
regex==2021.4.4
reportlab==3.5.67
requests==2.25.1
requests-oauthlib==1.3.0
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
team==1.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.7.4.3
udiskie==2.3.3
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xgboost==1.4.2
youtube-dl==2021.6.6
zipp==3.4.1
